localmind/backend/requirements.txt
fastapi
uvicorn[standard]
psutil
gputil
requests
llama-cpp-python
httpx
